Moving a few files over to Typescript and updating the bin/solidarity task.
