+++ bin/solidarity
@@ -4,7 +4,7 @@ const { isNewEnough, nodeMinimum } = require(`${__dirname}/../node_modules/glueg
 
 if (isNewEnough) {
   // let's kick it off with the command line parameters  
-  require(`${__dirname}/../src/index.js`)(process.argv)
+  require(`${__dirname}/../dist/index.js`)(process.argv)
 } else {
   console.log(
     "Whoops! You'll need to upgrade Node.js to go any further. \n\n" +
