--- src/commands/doctor.js
@@ -1,22 +1,0 @@
-// @solidarityDescription Check location of known apps.
-const run = async (context) => {
-  const { print, system } = context
-
-  const rn = await system.which(`react-native`)
-  const xcrun = await system.which('xcrun')
-  const androidHome = process.env['ANDROID_HOME']
-
-  // generic printing
-  print.table([
-    ['react native', rn],
-    ['xcrun', xcrun],
-    ['android home', androidHome]
-  ])
-}
-
-// Export command
-module.exports = {
-  description: '',
-  alias: 'd',
-  run
-}
