+++ tests/index.js
@@ -1,5 +1,5 @@
 import test from 'ava'
-import baseRuntimeConfiguration from '../src/index'
+import baseRuntimeConfiguration from '../dist/index'
 
 test('Verify Runtime AND Configuration', t => {
   t.snapshot(baseRuntimeConfiguration)
